refactor integration test module out
